move cache into own class

+++ lib/chuckle/curl.rb
@@ -3,38 +3,42 @@ require "ostruct"
 
 module Chuckle
   class Curl
-    def initialize(chuckle, request)
-      @chuckle = chuckle
+    def initialize(request)
       @request = request
     end
 
-    def tmp_body
-      @tmp_body ||= Util.tmp_path
+    # tmp path for response headers
+    def headers_path
+      @headers_path ||= Util.tmp_path
     end
 
-    def tmp_headers
-      @tmp_headers ||= Util.tmp_path
+    # tmp path for response body
+    def body_path
+      @body_path ||= Util.tmp_path
     end
 
-    # build the curl command line arguments
+    # the command line for this request, based on the request and the
+    # options from client
     def command(request)
+      client = request.client
       command = ["curl"]
       command << "--silent"
       command << "--compressed"
-      command += [ "--user-agent", @chuckle.user_agent]
-      command += ["--max-time", @chuckle.timeout]
-      command += ["--retry", @chuckle.nretries]
+      command += [ "--user-agent", client.user_agent]
+      command += ["--max-time", client.timeout]
+      command += ["--retry", client.nretries]
       command += ["--location", "--max-redirs", 3]
       if request.body
         command += ["--data-binary", request.body]
         command += ["--header", "Content-Type: application/x-www-form-urlencoded"]
       end
-      if @chuckle.cookie_jar
-        command += ["--cookie", @chuckle.cookie_jar]     # Read cookies from file
-        command += ["--cookie-jar", @chuckle.cookie_jar] # Write cookies to file
+      if client.cookie_jar
+        FileUtils.mkdir_p(File.dirname(cookie_jar))
+        command += ["--cookie", client.cookie_jar]     # Read cookies from file
+        command += ["--cookie-jar", client.cookie_jar] # Write cookies to file
       end
-      command += ["--dump-header", tmp_headers]
-      command += ["--output", tmp_body]
+      command += ["--dump-header", headers_path]
+      command += ["--output", body_path]
       command << request.uri
 
       command = command.map(&:to_s)
@@ -42,27 +46,32 @@ module Chuckle
     end
 
     def run
+      # note: explicitly use Kernel.system to allow for mocking
       command = command(@request)
-      # explicitly use Kernel to allow for mocking
       Kernel.system(*command)
 
       # capture exit code, bail on INT
       exit_code = $?.to_i / 256
-      if exit_code != 0
-        if $?.termsig == Signal.list["INT"]
-          Process.kill(:INT, $$)
-        end
+      if exit_code != 0 && $?.termsig == Signal.list["INT"]
+        Process.kill(:INT, $$)
       end
 
-      #
-      # fix tmp files if there were errors
-      #
-
+      # create tmp files if there were errors
+      if !File.exists?(body_path)
+        FileUtils.touch(body_path)
+      end
       if exit_code != 0
-        IO.write(tmp_body, "")
-        IO.write(tmp_headers, "exit_code #{exit_code}")
-      elsif !File.exists?(tmp_body)
-        FileUtils.touch(tmp_body)
+        IO.write(headers_path, Curl.exit_code_to_headers(exit_code))
+      end
+    end
+
+    def self.exit_code_to_headers(exit_code)
+      "exit_code #{exit_code}"
+    end
+
+    def self.exit_code_from_headers(headers)
+      if exit_code = headers[/^exit_code (\d+)/, 1]
+        exit_code.to_i
       end
     end
   end
